+++ ld64/src/ld/Options.cpp
@@ -1704,7 +1704,7 @@ void Options::addSection(const char* segment, const char* section, const char* p
 	::close(fd);
 
 	// record section to create
-	ExtraSection info = { segment, section, path, (uint8_t*)p, stat_buf.st_size };
+	ExtraSection info = { segment, section, path, (uint8_t*)p, (uint64_t) stat_buf.st_size };
 	fExtraSections.push_back(info);
 }
